python/2020-04-05_ShepBot/requirements.txt
discord.py
requests
praw
aiohttp